Support for fully static builds
Allows for compiling fully static builds, which can be useful
for fast deployment on systems like Alpine.

Now links with the mimalloc allocator by default instead of the
builtin. This is to avoid linking in non-glibc allocators on
platforms, like musl, which can significantly lower performance
on multithreaded systems.

Includes new packaging scripts for making binary releases when
new versions are pushed.