Add notify callback and entry access functions
This commit adds a new callback named 'notify' that can be used
to receive a stream of events for every change that occurs to
the cache. Each event include the new entry that was inserted or
updated, and the old entry that was deleted or replaced.

Along with the notify callback, there are four new functions for
accessing the key and value of an entry, and for retaining and
releasing an entry. These can be useful for operating on the
entries outside of the standard pogocache operations, such as
with secondary indexes, scripting, replication, etc.

The new functions are pogocache_entry_key, pogocache_entry_value,
pogocache_entry_retain, and pogocache_entry_release,