Added SCAN command
The new SCAN command works similar to the one
included with dragonfly and redis; where you
start with a zero cursor and continue making
requests using the next cursor returned by the
server, until the server returns zero.

What's different is that with Pogocache, the
number of returned keys will always be <= COUNT
and the scan does not block other clients.

Also the default COUNT is 100 instead of 10.

closes #32